requirements.txt
fastapi
uvicorn
pydantic
pandas==2.0.3
numpy==1.24.3
bs4
python-multipart
python-dotenv
slowapi
bcrypt>=3.1.7
python-jose[cryptography]